feat: Add script to add member to organization
This commit adds a new script `add_member.sh` that allows adding a member to the organization. The script prompts for the Epitech email of the user to be added and then uses GitHub API to send an invitation to the specified email. The script also includes placeholders for obtaining tokens for various services.

This script will help automate the process of adding members to the organization and improve the onboarding workflow.